requirements.txt
pandas==2.3.1
numpy==2.3.2
scipy==1.16.1
pyarrow==21.0.0
openpyxl==3.1.5
PyYAML==6.0.2
polars==1.32.2
bottleneck==1.5.0
numexpr==2.11.0
xgboost==3.0.3
scikit-learn==1.7.1